Renaming styled components
