+++ src/Components/Button/Button.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import DefaultButton from '../UI/Button/StyledButton';
+import DefaultButton from '../UI/Button/Button.styled';
 
 export default {
 	title: 'Components/Button',
