+++ src/Components/Product/ProductModeEdit.js
@@ -2,18 +2,18 @@ import Image from 'next/image';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useSWRConfig } from 'swr';
-import DefaultButton from '../UI/Button/StyledButton';
-import ButtonContainer from '../UI/Button/StyledButtonContainer';
-import FormContainer from '../UI/Form/StyledFormContainer';
-import FormElement from '../UI/Form/StyledFormElement';
-import Input from '../UI/Form/StyledInput';
-import InputContainer from '../UI/Form/StyledInputContainer';
-import InputFile from '../UI/Form/StyledInputFile';
-import InputSingleContainer from '../UI/Form/StyledInputSingleContainer';
-import Label from '../UI/Form/StyledLable';
-import LabelUpload from '../UI/Form/StyledLableUpload';
-import ImageContainer from '../UI/Image/StyledImageContainer';
-import ImageWrapper from '../UI/Image/StyledImageWrapper';
+import DefaultButton from '../UI/Button/Button.styled';
+import ButtonContainer from '../UI/Button/ButtonContainer.styled';
+import FormContainer from '../UI/Form/FormContainer.styled';
+import FormElement from '../UI/Form/FormElement.styled';
+import Input from '../UI/Form/Input.styled';
+import InputContainer from '../UI/Form/InputContainer.styled';
+import InputFile from '../UI/Form/InputFile.styled';
+import InputSingleContainer from '../UI/Form/InputSingleContainer.styled';
+import Label from '../UI/Form/Lable.styled';
+import LabelUpload from '../UI/Form/LableUpload.styled';
+import ImageContainer from '../UI/Image/ImageContainer.styled';
+import ImageWrapper from '../UI/Image/ImageWrapper.styled';
 import Typography from '../UI/Typography';
 
 export default function ProductModeEdit({
